NaNStatistics

NaNStatistics.inpctile

NaNStatistics.movmean

NaNStatistics.nanaad

NaNStatistics.nanadd

NaNStatistics.nanadd!

NaNStatistics.nanbinmean

NaNStatistics.nanbinmean

NaNStatistics.nanbinmean!

NaNStatistics.nanbinmean!

NaNStatistics.nanbinmedian

NaNStatistics.nanbinmedian!

NaNStatistics.nanbinwmean

NaNStatistics.nanextrema

NaNStatistics.nanmad

NaNStatistics.nanmask

NaNStatistics.nanmask!

NaNStatistics.nanmax

NaNStatistics.nanmaximum

NaNStatistics.nanmean

NaNStatistics.nanmedian

NaNStatistics.nanmin

NaNStatistics.nanminimum

NaNStatistics.nanpctile

NaNStatistics.nanrange

NaNStatistics.nanstd

NaNStatistics.nansum

NaNStatistics.zeronan!

NaNStatistics.inpctile — Method

inpctile(A, p::Number; dims)

Return a boolean array that identifies which values of the iterable collection A fall within the central pth percentile, optionally along a dimension specified by dims.

A valid percentile value must satisfy 0 <= p <= 100.

NaNStatistics.movmean — Method

movmean(x::AbstractVecOrMat, n::Number)

Simple moving average of x in 1 or 2 dimensions, spanning n bins (or n*n in 2D), returning an array of the same size as x. For the resulting moving average to be symmetric, n must be an odd integer; if n is not an odd integer, the first odd integer greater than n will be used instead.

NaNStatistics.nanaad — Method

nanaad(A; dims)

Mean (average) absolute deviation from the mean, ignoring NaNs, of an indexable collection A, optionally along a dimension specified by dims. Note that for a Normal distribution, sigma = 1.253 * AAD

NaNStatistics.nanadd! — Method

nanadd!(A, B)

Add the non-NaN elements of B to A, treating NaNs as zeros

NaNStatistics.nanadd — Method

nanadd(A, B)

Add the non-NaN elements of A and B, treating NaNs as zeros

NaNStatistics.nanbinmean! — Method

nanbinmean!(MU, N, x, y, z, xedges::AbstractRange, yedges::AbstractRange)

Ignoring NaNs, fill the matrix MU with the means and N with the counts of non-NAN z values that fall into a 2D grid of x and y bins defined by xedges and yedges. The independent variables x and y, as well as the dependent variable z, are all expected as 1D vectors (any subtype of AbstractVector).

The output matrices MU and N must be the same size, and must each have length(yedges)-1 rows and length(xedges)-1 columns.

NaNStatistics.nanbinmean! — Method

nanbinmean!(MU, [N], x, y, xedges::AbstractRange)

Ignoring NaNs, fill the array MU with the means (and optionally N with the counts) of non-NAN y values that fall into each of length(xedges)-1 equally spaced bins along the x axis with bin edges specified by xedges.

The array of x data should given as a one-dimensional array (any subtype of AbstractVector) and y as either a 1-d or 2-d array (any subtype of AbstractVecOrMat).

The output arrays MU and N must be the same size, and must have the same number of columns as y; if y is a 2-d array (matrix), then each column of y will be treated as a separate variable.

NaNStatistics.nanbinmean — Method

nanbinmean(x, y, xedges::AbstractRange)

Ignoring NaNs, calculate the mean of y values that fall into each of N = length(xedges)-1 equally spaced bins along the x axis with bin edges specified by xedges.

The array of x data should given as a one-dimensional array (any subtype of AbstractVector) and y as either a 1-d or 2-d array (any subtype of AbstractVecOrMat). If y is a 2-d array, then each column of y will be treated as a separate variable.

NaNStatistics.nanbinmean — Method

nanbinmean(x, y, z, xedges, yedges)

Ignoring NaNs, calculate the mean of z values that fall into a 2D grid of x and y bins with bin edges defined by xedges and yedges. The independent variables x and y, as well as the dependent variable z, are all expected as 1D vectors (any subtype of AbstractVector).

NaNStatistics.nanbinmedian! — Method

nanbinmedian!(M, [N], x, y, xedges::AbstractRange)

Fill the array M with the medians (and optionally N with the counts) of non-NaN y values that fall into each of length(xedges)-1 equally spaced bins along the x axis with bin edges specified by xedges.

If y is a 2-d array (matrix), each column will be treated as a separate variable

NaNStatistics.nanbinmedian — Method

nanbinmedian(x, y, xedges::AbstractRange)

Calculate the median, ignoring NaNs, of y values that fall into each of length(xedges)-1 equally spaced bins along the x axis with bin edges specified by xedges.

If y is a 2-d array (matrix), each column will be treated as a separate variable

NaNStatistics.nanbinwmean — Method

nanbinwmean(x, y, xedges::AbstractRange)

Ignoring NaNs, calculate the weighted mean of y values that fall into each of length(xedges)-1 equally spaced bins along the x axis with bin edges specified by xedges.

The array of x data should given as a one-dimensional array (any subtype of AbstractVector) and y as either a 1-d or 2-d array (any subtype of AbstractVecOrMat). If y is a 2-d array, then each column of y will be treated as a separate variable.

NaNStatistics.nanextrema — Method

nanextrema(A; dims)

Find the extrema (maximum & minimum) of an indexable collection A, ignoring NaNs, optionally along a dimension specified by dims.

NaNStatistics.nanmad — Method

nanmad(A; dims)

Median absolute deviation from the median, ignoring NaNs, of an indexable collection A, optionally along a dimension specified by dims. Note that for a Normal distribution, sigma = 1.4826 * MAD

NaNStatistics.nanmask! — Method

nanmask!(mask, A)

Fill a Boolean mask of dimensions size(A) that is false wherever A is NaN

NaNStatistics.nanmask — Method

nanmask(A)

Create a Boolean mask of dimensions size(A) that is false wherever A is NaN

NaNStatistics.nanmax — Method

nanmax(a,b)

As max(a,b), but if either argument is NaN, return the other one

NaNStatistics.nanmaximum — Method

nanmaximum(A; dims)

Find the largest non-NaN value of an indexable collection A, optionally along a dimension specified by dims.

NaNStatistics.nanmean — Method

nanmean(A, [W]; dims)

Ignoring NaNs, calculate the mean (optionally weighted) of an indexable collection A, optionally along dimensions specified by dims.

NaNStatistics.nanmedian — Method

nanmedian(A; dims)

Calculate the median, ignoring NaNs, of an indexable collection A, optionally along a dimension specified by dims.

NaNStatistics.nanmin — Method

nanmin(a,b)

As min(a,b), but if either argument is NaN, return the other one

NaNStatistics.nanminimum — Method

nanminimum(A; dims)

As minimum but ignoring NaNs: Find the smallest non-NaN value of an indexable collection A, optionally along a dimension specified by dims.

NaNStatistics.nanpctile — Method

nanpctile(A, p; dims)

Find the pth percentile of an indexable collection A, ignoring NaNs, optionally along a dimension specified by dims.

A valid percentile value must satisfy 0 <= p <= 100.

NaNStatistics.nanrange — Method

nanrange(A; dims)

Calculate the range (maximum - minimum) of an indexable collection A, ignoring NaNs, optionally along a dimension specified by dims.

NaNStatistics.nanstd — Method

nanstd(A, [W]; dims)

Calculate the standard deviation (optionaly weighted), ignoring NaNs, of an indexable collection A, optionally along a dimension specified by dims.

NaNStatistics.nansum — Method

nansum(A; dims)

Calculate the sum of an indexable collection A, ignoring NaNs, optionally along dimensions specified by dims.

NaNStatistics.zeronan! — Method

zeronan!(A)

Replace all NaNs in A with zeros of the same type